Fix: ListHeading state issue ok
